+++ crabpy_pyramid/routes/crab.py
@@ -23,6 +23,7 @@ def includeme(config):
     config.add_route('get_huisnummer_by_id', '/crab/huisnummers/{huisnummer_id}')
     config.add_route('list_percelen', '/crab/huisnummers/{huisnummer_id}/percelen')
     config.add_route('get_perceel_by_id', '/crab/percelen/{perceel_id1}/{perceel_id2}')
+    config.add_route('list_huisnummers_by_perceel', '/crab/percelen/{perceel_id1}/{perceel_id2}/huisnummers')
     config.add_route('list_gebouwen', '/crab/huisnummers/{huisnummer_id}/gebouwen')
     config.add_route('get_gebouw_by_id', '/crab/gebouwen/{gebouw_id}')
     config.add_route('list_subadressen', '/crab/huisnummers/{huisnummer_id}/subadressen')