+++ CHANGES.rst
@@ -4,6 +4,8 @@
 - Update dependencies. Update pycountry to version 1.19.
 - Add explicit support for python 3.5.
 - Add deelgemeenten. (#33)
+- Added postadressen in a few places.
+- Added a list of huisnummers to a CRAB perceel.
 
 0.3.0 (2015-06-01)
 ------------------